Added notification from server to client regarding job progress

+++ framework/bundles/org.jboss.tools.rsp.launching/src/main/java/org/jboss/tools/rsp/launching/utils/SimpleProgressMonitor.java
@@ -69,7 +69,7 @@ public class SimpleProgressMonitor implements IProgressMonitor {
 			currWork = 0;
 	}
 
-	public int getPercentage() {
+	public double getPercentage() {
 		if (done) {
 			return 100;
 		}
